requirements.txt
discord.py==1.7.3
python-dotenv==0.20.0
praw==7.6.0
openai==0.18.1
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1